fix(collector): reduce time window for non-monotonicity, add timing grace window (#234)
When message timestamps are collected a lot in advance of sending the message, userspace can decide to process a message that will be later (because it will see that message first).

move the timestamp computation as close as we could to sending the perf measurement so the race window is reduced

add an explicit grace period before current time, where userspace will not read messages that are closer than that period, in case a message took that long to send. If all messages take less than the grace period to send, they will be processed in order.

* move timestamps close to sending messages, to reduce time until observed by userspace

* cargo fmt dispatcher.rs

* add grace period (currently 100 microseconds) to avoid races with events